refactor: split up modules to avoid circular import
Put `ensure_package_and_question_state_exist` into separate
module`decorators.py`.
